+++ public/App.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Blog;
+namespace Blog\public;
 
 use \Blog\controller\Config;
 use \Blog\model\DBFactory;
@@ -56,21 +56,5 @@ class App
         $_COOKIE['ketto'] = self::$ketto;
 
         $_SESSION['ketto'] = self::$ketto;
-    }
-
-    public static function newTurn()
-    {
-        if ($_COOKIE['ketto'] == $_SESSION['ketto']) {
-            // C'est reparti pour un tour
-            self::$ketto = session_id().microtime().rand(0,9999999999);
-            self::$ketto = hash('sha512', self::$ketto);
-            $_COOKIE['ketto'] = self::$ketto;
-            $_SESSION['ketto'] = self::$ketto;
-        } else {
-            // On détruit la session
-            $_SESSION = [];
-            session_destroy();
-            header('location:index.php');
-        }
     }
 }
